requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
urllib3>=2.0.0
black
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
